xpserver/requirements.txt
flask>=0.10
flask-sqlalchemy>=1
gevent
gevent-websocket